pollProject/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.4
django-bootstrap4==23.2
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
